Bug correction + creation of grid
In the past, modifying the map, the tile format or whatever else variable from the UI resulted in a bug. This has been addressed. Now users can even process multiple maps at once.

On the UI side, I added gridlines on the map, so that users can have an idea of how their map will be processed according to different tile formats.