chunks lengths bug correction
Setting an arbitrary number (300px) for the size of chunks turned out to be a bad idea. Now the number has been changed to 384, which can be divided into an integer using any of the tile formats available.
In case some more trouble happens with the cutting of the map into chunks I added one more security layer in the form of a condition: if somehow the chunks can't be divided in an integer number of tiles, give up on cutting the map into chunks altogether.